Enqueue incoming requests for rate-limited processing.
Mike Sampson said his Liftingcast will have problems when changes are made in
rapid succession because the CouchDB changes feed (source of data to LC) batches
multiple changes to the same document that occur within some window of time.

Handling one request per second seems to work fine.

We put requests on a channel in the Ring handler and return 202 immediately so
that we don't block DRL. (A message queue would be a better solution but this
approach doesn't require Scott to troubleshoot Redis on the Pi when he has
little time away from working the meet.)