Fix changes in entity specs to allow for changes I haven't made.
I don't make all of the changes to the documents. Ex: The scorekeeper enters the
weight for the next attempt. Therefore, an attempt's changes will include a
change(s) to the weight. My specs should not disallow that!

On the other hand, I want to ensure that I don't make any unplanned changes to
documents.

Therefore, I apply entity-specific specs to my changes before I add them to a
document while I check that the changes field of an entity document with a more
permissive spec.