Add command-line arguments and logging.
Adding command-line arguments required removing the defs in the namespace and
"initing" many values (e.g., db) dynamically, in response to running main,
instead of statically, with `def`s anywhere in the namespace.

Logging verbosity was a command line argument so I added Timbre to ease that
change.

This commit should be broken up into many small commits but it's too late and
I'm too tired and this must be finished.