Remove logging of non-platform changes.
I was logging changes to
* the platform
* referees
* attempts
* other changes

I previously had the attempt ID spec checking against a master list of attempt
IDs for the platform. Similarly for referee IDs. I removed those checks when I
switched the code from deffing lots of stuff to initing in -main.

The only thing from the whole changes feed that I'm acting on is changes to the
current attempt ID for the platform; I don't need the rest.