Reduce delay before lights come on.
I was
* Getting all docs from the database (blocking request)
    * And then turning the lights on
        * And then turning the lights off after the given duration
    * Concurrently, waiting a short time and then marking the attempt and
      selecting the next attempt

The lights, therefore, couldn't come on until the whole database had been
fetched and indexed.

Instead, I
* Get only the referee docs and update them to turn the lights on
    * And then turn the lights off after the given duration
* Concurrently get all docs
    * And then mark the attempt and select the next attempt

The lights come on more quickly now, since we're fetching 3 documents instead of
on the order of 1000.

Marking the attempt and selecting the next attempt is naturally delayed by the
blocking request to get all docs from the database, which are required to find
the next attempt. There is no need for an artificial delay to ensure that the
lights come on first.