Allow next attempt id to be nil (i.e., end of session)
In Liftingcast, the current attempt id of a platform is set to nil when there
are no more attempts with weights and without results. I.e., at the end of a
session.
My specs were not allowing the current attempt id of a platform to be nil and
were not allowing nil as the value of a change to the current attempt id of a
platform.
When Liftingcast reached the end of the round, my code was marking the attempt
but not updating the platform so the attempt was still selected in the
Liftingcast GUI and the platform state was out of sync with reality.

Making attempt id nilable in the appropriate specs fixes this.