Separate resetting clock from selecting next attempt
While it would be nice to have an atomic UI update that includes both of these
actions, the UI updates on the client happen as document changes come in -- they
are not atomic.

Resetting the clock requires no computation; it's very fast.
Selecting the next attempt requires pulling the entire meet database and
computing the next lifter; it's relatively slow.

Therefore, let's reset the clock "immediately" after an attempt is judged so
that the UI looks as correct and up-to-date as possible for the users. Then
we'll select the next attempt as soon as we can.